+++ examples/with-thirdparty/api-server.js
@@ -58,9 +58,7 @@ supertokens.init({
             },
         }),
         Session.init(),
-        Dashboard.init({
-            apiKey: "supertokens_is_awesome",
-        }),
+        Dashboard.init(),
     ],
 });